PR-template
This is the template for PRs. It is just a markdown file. I am committing it on a new branch (as opposed to pushing it directly to master), meaning it will need its own pull request approved. There is no need to fetch the branch and look at it because it is just a markdown file. You will have to approve it and when it is merged to master every new pull request will auto-populate with this text in the pull request text field.